refactor: move DownloadTask and DeleteTask to org.terasology.launcher.tasks package
