+++ src/main/java/org/terasology/launcher/gui/javafx/ApplicationController.java
@@ -49,6 +49,8 @@ import javafx.stage.StageStyle;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.terasology.launcher.game.GameStarter;
+import org.terasology.launcher.tasks.DeleteTask;
+import org.terasology.launcher.tasks.DownloadTask;
 import org.terasology.launcher.packages.Package;
 import org.terasology.launcher.packages.PackageManager;
 import org.terasology.launcher.settings.BaseLauncherSettings;
