+++ src/main/java/org/terasology/launcher/packages/PackageManager.java
@@ -33,7 +33,7 @@ import org.terasology.launcher.packages.db.DatabaseRepositoryDeserializer;
 import org.terasology.launcher.util.DownloadException;
 import org.terasology.launcher.util.DownloadUtils;
 import org.terasology.launcher.util.FileUtils;
-import org.terasology.launcher.util.ProgressListener;
+import org.terasology.launcher.tasks.ProgressListener;
 
 import java.io.File;
 import java.io.IOException;
