+++ src/main/java/org/terasology/launcher/tasks/ProgressListener.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.terasology.launcher.util;
+package org.terasology.launcher.tasks;
 
 public interface ProgressListener {
