+++ src/main/java/org/terasology/launcher/util/DownloadUtils.java
@@ -19,6 +19,7 @@ package org.terasology.launcher.util;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.terasology.launcher.game.GameJob;
+import org.terasology.launcher.tasks.ProgressListener;
 import org.w3c.dom.Document;
 import org.w3c.dom.Node;
 import org.w3c.dom.NodeList;
